Migrate skill search to slash commands Unify skill search with character search
