+++ DanMemoDiscordBot/mypy.ini
